+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 - Unreleased
     - Added Model Orchestra meta estimator
+    - Added Stop Word Filter transformer
     - Added document frequency smoothing to TF-IDF Transformer
     - Added Uniform neural net weight initializer
     - Fixed missing probabilities in Classification Tree